Adams' Thesis data

Data from a replication of Douven and Verbrugge (2010) investigating the relationship between acceptability ratings of conditional sentences and conditional probabilities (variable rating) across three different types of conditionals (inductive, abductive, deductive).

Usage

AT

Format

A data frame with 4,590 rows and 5 variables:

submission_id

A unique identifier for each participant.

item_nr

An identifier for the context-statement pair presented to the participant.

rating

The kind of rating the participant has given.

cond_type

The type of conditional statement.

response

The measured acceptability rating on a 7-point Likert scale.

References

Douven, I., Verbrugge, S. (2010). The Adams Family. Cognition, 117, 302-18. https://doi.org/10.1016/j.cognition.2010.08.015.

Examples

if (FALSE) {
# fit an ordinal model using 'brms'
# regress response as a function of rating and conditional type (with interaction)
fit <- brms::brm(response ~ rating * cond_type,
                 data = AT,
                 family = cumulative("logit"))

# check encoding scheme used in the model
get_cell_definitions(fit)

# get draws for acceptability averaged over the type of conditional
extract_cell_draws(fit, rating == "acceptability")

# compare main effects of rating
compare_groups(fit, rating == "acceptability", rating == "cond_probability")
}